feat: :sparkles: Created new repositories, controllers, routers
Created controllers, repositories and router for Categoria, Produto, Lote, changed some columns types

+++ src/app/models/Produto.ts
@@ -17,7 +17,7 @@ class Produto {
     @Column('varchar', { length: 150, nullable: false })
     descricao: string;
 
-    @Column('int', { nullable: false })
+    @Column('float', { nullable: false })
     custo: number;
 
     @Column('float', { nullable: false })
